Move shorthand background parsing to this package
