+++ packages/css-data/package.json
@@ -21,7 +21,6 @@
     "@webstudio-is/jest-config": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
     "camelcase": "^6.3.0",
-    "css-tree": "^2.3.1",
     "html-tags": "^3.2.0",
     "jest": "^29.3.1",
     "mdn-data": "2.0.30",
@@ -50,6 +49,7 @@
   "dependencies": {
     "colord": "^2.9.3",
     "hyphenate-style-name": "^1.0.4",
-    "warn-once": "^0.1.1"
+    "warn-once": "^0.1.1",
+    "css-tree": "^2.3.1"
   }
 }
