+++ packages/css-data/src/parse-css.ts
@@ -1,6 +1,8 @@
 import * as csstree from "css-tree";
-import { parseCssValue } from "./parse-css-value";
-import { type StyleProperty, StyleValue } from "./schema";
+import { parseCssValue as parseCssValueLonghand } from "./parse-css-value";
+import { parseBackground } from "./property-parsers/background";
+import { properties } from "./__generated__/properties";
+import { type StyleProperty, type Style as S, StyleValue } from "./schema";
 
 type Selector = string;
 type Style = {
@@ -10,6 +12,24 @@ type Style = {
 };
 
 export type Styles = Record<Selector, Style[]>;
+
+const parseCssValue = function parseCssValue(
+  property: string,
+  value: string
+): S | null {
+  if (property === "background") {
+    return parseBackground(value);
+  }
+
+  if (property in properties) {
+    return {
+      [property]: parseCssValueLonghand(property as StyleProperty, value),
+    };
+  }
+
+  return null;
+};
+
 export const parseCss = function cssToWS(css: string) {
   const ast = csstree.parse(css);
 
@@ -23,39 +43,49 @@ export const parseCss = function cssToWS(css: string) {
     if (node.type === "ClassSelector") {
       if (!item.prev && !item.next) {
         selectors.push(node.name);
-        return;
       }
       return;
     }
 
     if (node.type === "Declaration") {
-      const property = node.property as StyleProperty;
       const stringValue = csstree.generate(node.value);
-      const value = parseCssValue(property, stringValue);
-      try {
-        StyleValue.parse(value);
-        selectors.forEach((selector) => {
-          if (Array.isArray(styles[selector])) {
-            styles[selector].push({ property, value });
-          } else {
-            styles[selector] = [{ property, value }];
-          }
-        });
-      } catch (error) {
-        if (process.env.NODE_ENV !== "production") {
-          // eslint-disable-next-line no-console
-          console.warn(
-            true,
-            `Declaration parsing for \`${selectors.join(
-              ", "
-            )}.${property}: ${stringValue}\` failed:\n\n${JSON.stringify(
-              value,
-              null,
-              2
-            )}`
-          );
-        }
+      const parsedCss = parseCssValue(node.property, stringValue);
+
+      if (!parsedCss) {
+        return;
       }
+
+      (Object.entries(parsedCss) as [StyleProperty, StyleValue][]).forEach(
+        ([property, value]) => {
+          try {
+            StyleValue.parse(value);
+            selectors.forEach((selector) => {
+              if (Array.isArray(styles[selector])) {
+                styles[selector].push({
+                  property,
+                  value,
+                });
+              } else {
+                styles[selector] = [{ property, value }];
+              }
+            });
+          } catch (error) {
+            if (process.env.NODE_ENV !== "production") {
+              // eslint-disable-next-line no-console
+              console.warn(
+                true,
+                `Declaration parsing for \`${selectors.join(
+                  ", "
+                )}.${property}: ${stringValue}\` failed:\n\n${JSON.stringify(
+                  value,
+                  null,
+                  2
+                )}`
+              );
+            }
+          }
+        }
+      );
     }
   });
