+++ packages/css-data/src/property-parsers/index.ts
@@ -1,0 +1,1 @@
+export * from "./background";
